Move the binary sensors onto the binary_sensor platform
Online, Enabled and Out Of Date displayed as raw "on"/"off". The cause was not the
labels: these were BinarySensorEntity subclasses handed to the sensor platform, and
Home Assistant derives an entity domain from the platform that creates it, not from
the class. They therefore lived in the sensor domain, where the binary-sensor device
class wording is never applied.

In their own platform the same entities read as Connected/Disconnected, OK/Problem
and Running/Not running, supplied by Home Assistant from the device class with no
per-entity strings to maintain. All sixteen moved, not only the three proxy ones
that prompted this: repository online, out-of-date, immutable, accessible and the
two capacity warnings, server health and connectivity, license auto-update and
cloud connect, and the three HA cluster sensors.

Two consequences worth knowing.

The entity IDs move from sensor.* to binary_sensor.*. Unique IDs are unchanged, so
Home Assistant would otherwise keep both — the old one stranded and unavailable
forever, since nothing provides it. Each replacement removes its predecessor by
unique ID and logs it, so one upgrade migrates rather than duplicates. Automations
and dashboards referencing the old IDs need updating; the README says so.

It also fixes three shipped blueprints. proxy_offline and ha_cluster_failover filter
their entity selectors on domain: binary_sensor, which matched nothing at all while
these entities were in the sensor domain.

Proxy Enabled keeps a custom icon rather than a device class, because there is no
enabled/disabled class and PROBLEM would colour a deliberately disabled proxy red.
